newsletter/requirements.txt
psycopg2-binary
boto3
python-dotenv
xhtml2pdf
